remove getblocktemplate trait
